refactor: add packages to dto
