+++ BE/src/main/java/com/codesquad/issue04/service/LabelService.java
@@ -8,9 +8,9 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.codesquad.issue04.domain.label.Label;
 import com.codesquad.issue04.domain.label.LabelRepository;
-import com.codesquad.issue04.web.dto.request.LabelCreateRequestDto;
-import com.codesquad.issue04.web.dto.request.LabelDeleteRequestDto;
-import com.codesquad.issue04.web.dto.request.LabelUpdateRequestDto;
+import com.codesquad.issue04.web.dto.request.label.LabelCreateRequestDto;
+import com.codesquad.issue04.web.dto.request.label.LabelDeleteRequestDto;
+import com.codesquad.issue04.web.dto.request.label.LabelUpdateRequestDto;
 import com.codesquad.issue04.web.dto.response.label.LabelDetailResponseDto;
 import com.codesquad.issue04.web.dto.response.label.LabelOverviewDto;
 import com.codesquad.issue04.web.dto.response.label.LabelOverviewResponseDtos;
