+++ BE/src/main/java/com/codesquad/issue04/web/dto/request/issue/IssueCreateRequestDto.java
@@ -1,4 +1,4 @@
-package com.codesquad.issue04.web.dto.request;
+package com.codesquad.issue04.web.dto.request.issue;
 
 import java.util.Collections;
 import java.util.List;
