+++ BE/src/main/java/com/codesquad/issue04/web/dto/request/label/LabelUpdateRequestDtoTemporary.java
@@ -1,4 +1,4 @@
-package com.codesquad.issue04.web.dto.request;
+package com.codesquad.issue04.web.dto.request.label;
 
 import lombok.Getter;
