+++ BE/src/test/java/com/codesquad/issue04/service/IssueServiceTest.java
@@ -24,13 +24,13 @@ import com.codesquad.issue04.domain.issue.vo.Status;
 import com.codesquad.issue04.domain.milestone.Milestone;
 import com.codesquad.issue04.domain.milestone.NullMilestone;
 import com.codesquad.issue04.domain.user.RealUser;
-import com.codesquad.issue04.web.dto.request.CommentCreateRequestDto;
-import com.codesquad.issue04.web.dto.request.CommentDeleteRequestDto;
-import com.codesquad.issue04.web.dto.request.CommentUpdateRequestDto;
-import com.codesquad.issue04.web.dto.request.IssueCloseRequestDto;
-import com.codesquad.issue04.web.dto.request.IssueCreateRequestDto;
-import com.codesquad.issue04.web.dto.request.IssueDeleteRequestDto;
-import com.codesquad.issue04.web.dto.request.IssueUpdateRequestDto;
+import com.codesquad.issue04.web.dto.request.comment.CommentCreateRequestDto;
+import com.codesquad.issue04.web.dto.request.comment.CommentDeleteRequestDto;
+import com.codesquad.issue04.web.dto.request.comment.CommentUpdateRequestDto;
+import com.codesquad.issue04.web.dto.request.issue.IssueCloseRequestDto;
+import com.codesquad.issue04.web.dto.request.issue.IssueCreateRequestDto;
+import com.codesquad.issue04.web.dto.request.issue.IssueDeleteRequestDto;
+import com.codesquad.issue04.web.dto.request.issue.IssueUpdateRequestDto;
 import com.codesquad.issue04.web.dto.response.error.ErrorResponseDto;
 import com.codesquad.issue04.web.dto.response.issue.IssueDetailResponseDto;
 import com.codesquad.issue04.web.dto.response.issue.IssueOverviewDto;
